#14 feat: adiciona tratamentos de erro em mitologia

+++ src/main/java/riordanverse/riordanverse/services/MitologiaService.java
@@ -20,7 +20,7 @@ public class MitologiaService {
 		// Verifica se a mitologia com o nome já existe
 		Mitologia existente = mitologiaRepository.findByNome(nome);
 		if (existente != null) {
-			throw new IllegalStateException("Já existe uma mitologia com o nome: " + nome);
+			throw new IllegalArgumentException("Já existe uma mitologia com o nome: " + nome);
 		}
 
 		// Cria uma nova mitologia se não existir
@@ -54,14 +54,38 @@ public class MitologiaService {
 	}
 
 	public Mitologia salvar(Mitologia mitologia) {
+		String nome = mitologia.getNome();
+		Mitologia existente = mitologiaRepository.findByNome(nome);
+
+		if (existente != null) {
+			throw new RuntimeException("Já existe uma mitologia com o nome: " + nome);
+		}
+
 		return mitologiaRepository.save(mitologia);
 	}
 
 	public Mitologia atualizar(Mitologia mitologia) {
+		String nome = mitologia.getNome();
+		Mitologia existente = getMitologiaByNome(nome);
+
+		if (existente != null) {
+			throw new RuntimeException("Já existe uma mitologia com o nome: " + nome);
+		}
+
+		if (mitologia.getId() == null) {
+			throw new RuntimeException("É necessário informar o id da mitologia a ser atualizada.");
+		}
+
 		return mitologiaRepository.save(mitologia);
 	}
 
 	public void remover(Integer idMitologia) {
+		Mitologia existente = getMitologiaById(idMitologia);
+
+		if (existente == null) {
+			throw new RuntimeException("Não foi encontrada uma mitologia com esse id.");
+		}
+
 		mitologiaRepository.deleteById(idMitologia);
 	}
